assumptions: Move _ask_recursive to ask.py
